Isolate a listener's failure, and pin the marker that made it safe
The branch below marks an error escaping host code so an outer exchange
never recovers from it — that is what stops a call the server already
executed from being sent twice. This branch's notification routing
isolates a listener's failure outright, so in practice such an error no
longer reaches an outer recovery at all, and its examples now pin the
wire (the executed call is sent once) rather than the propagation. The
marker itself gets a direct example, since the routing would otherwise be
the only thing under test.

The stdio restart example waits on the exit being handled, not on the
lazy retirement this branch pre-empts, and the input round-trip
fulfilment moves into its own mixin to keep the shared module under its
length cop.
